Move all files to a separate namespace.
