+++ src/FullFACPreconditioner.cpp
@@ -1,17 +1,6 @@
-// ---------------------------------------------------------------------
-//
-// Copyright (c) 2014 - 2020 by the IBAMR developers
-// All rights reserved.
-//
-// This file is part of IBAMR.
-//
-// IBAMR is free software and is distributed under the 3-clause BSD
-// license. The full text of the license can be found in the file
-// COPYRIGHT at the top level directory of IBAMR.
-//
-// ---------------------------------------------------------------------
-
 /////////////////////////////// INCLUDES /////////////////////////////////////
+
+#include "multiphase/FullFACPreconditioner.h"
 
 #include "ibtk/CartCellDoubleCubicCoarsen.h"
 #include "ibtk/CartSideDoubleCubicCoarsen.h"
@@ -23,7 +12,6 @@
 #include "ibtk/ibtk_enums.h"
 #include "ibtk/namespaces.h" // IWYU pragma: keep
 
-#include "FullFACPreconditioner.h"
 #include "MultiblockDataTranslator.h"
 #include "PatchHierarchy.h"
 #include "SAMRAIVectorReal.h"
@@ -43,7 +31,7 @@ static Timer* t_initialize;
 static Timer* t_deallocate;
 } // namespace
 
-namespace IBTK
+namespace multiphase
 {
 /////////////////////////////// STATIC ///////////////////////////////////////
