+++ time_stepping/time_stepping.cpp
@@ -1,15 +1,4 @@
-// ---------------------------------------------------------------------
-//
-// Copyright (c) 2017 - 2020 by the IBAMR developers
-// All rights reserved.
-//
-// This file is part of IBAMR.
-//
-// IBAMR is free software and is distributed under the 3-clause BSD
-// license. The full text of the license can be found in the file
-// COPYRIGHT at the top level directory of IBAMR.
-//
-// ---------------------------------------------------------------------
+#include "multiphase/INSVCTwoFluidStaggeredHierarchyIntegrator.h"
 
 #include <ibamr/StaggeredStokesSolverManager.h>
 #include <ibamr/StokesSpecifications.h>
@@ -33,8 +22,7 @@
 #include <SAMRAI_config.h>
 #include <StandardTagAndInitialize.h>
 
-// Local includes
-#include "INSVCTwoFluidStaggeredHierarchyIntegrator.h"
+using namespace multiphase;
 
 /*******************************************************************************
  * For each run, the input filename must be given on the command line.  In all *
